int-ext-resolv-mapper/requirements.txt
pyasn
dnslib
ripe.atlas.cousteau
